fix: Added new structure to Instagram icon
