+++ src/components/index.ts
@@ -149,7 +149,9 @@ export * from './social-icons/Reddit';
 export * from './social-icons/Twitter';
 export * from './social-icons/Facebook';
 export * from './social-icons/Linkedin';
-export * from './social-icons/Instagram';
+export * from './social-icons/Instagram/Instagram';
+export * from './social-icons/Instagram/Instagram18';
+export * from './social-icons/Instagram/Instagram24';
 export * from './social-icons/Wikipedia';
 export * from './social-icons/Youtube';
 export * from './social-icons/Docs/Docs';
