--- src/components/social-icons/Instagram.tsx
@@ -1,54 +1,0 @@
-import React, { FC } from 'react';
-import { ISocialIconProps } from './type';
-
-export const IconInstagram: FC<ISocialIconProps> = ({
-	size = 16,
-	color = 'currentColor',
-}) => {
-	const size18 = (
-		<svg
-			width='18'
-			height='18'
-			viewBox='0 0 18 18'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M12.2811 2.43738C13.1506 2.43998 13.9838 2.78654 14.5986 3.40138C15.2135 4.01621 15.56 4.84936 15.5626 5.71887V12.2811C15.56 13.1506 15.2135 13.9838 14.5986 14.5986C13.9838 15.2135 13.1506 15.56 12.2811 15.5626H5.71887C4.84936 15.56 4.01621 15.2135 3.40138 14.5986C2.78654 13.9838 2.43998 13.1506 2.43738 12.2811V5.71887C2.43998 4.84936 2.78654 4.01621 3.40138 3.40138C4.01621 2.78654 4.84936 2.43998 5.71887 2.43738H12.2811ZM12.2811 1.125H5.71887C3.19219 1.125 1.125 3.19219 1.125 5.71887V12.2811C1.125 14.8078 3.19219 16.875 5.71887 16.875H12.2811C14.8078 16.875 16.875 14.8078 16.875 12.2811V5.71887C16.875 3.19219 14.8078 1.125 12.2811 1.125Z'
-				fill={color}
-			/>
-			<path
-				d='M13.2655 5.71887C13.0708 5.71887 12.8805 5.66114 12.7186 5.55297C12.5567 5.44481 12.4306 5.29107 12.3561 5.1112C12.2816 4.93133 12.2621 4.7334 12.3 4.54245C12.338 4.3515 12.4318 4.1761 12.5694 4.03843C12.7071 3.90077 12.8825 3.80701 13.0735 3.76903C13.2644 3.73105 13.4623 3.75054 13.6422 3.82505C13.8221 3.89955 13.9758 4.02572 14.084 4.1876C14.1921 4.34948 14.2499 4.5398 14.2499 4.73449C14.2501 4.86384 14.2249 4.99197 14.1755 5.11153C14.1261 5.23108 14.0536 5.33971 13.9622 5.43117C13.8707 5.52263 13.7621 5.59513 13.6425 5.6445C13.523 5.69387 13.3948 5.71915 13.2655 5.71887ZM9 6.37488C9.51919 6.37488 10.0267 6.52885 10.4584 6.8173C10.8901 7.10575 11.2266 7.51574 11.4253 7.99541C11.624 8.47509 11.676 9.00291 11.5747 9.51214C11.4734 10.0214 11.2234 10.4891 10.8562 10.8562C10.4891 11.2234 10.0214 11.4734 9.51213 11.5747C9.00291 11.676 8.47509 11.624 7.99541 11.4253C7.51573 11.2266 7.10574 10.8901 6.81729 10.4584C6.52884 10.0267 6.37488 9.5192 6.37488 9C6.37563 8.30401 6.65244 7.63673 7.14458 7.14459C7.63672 6.65244 8.304 6.37563 9 6.37488ZM9 5.0625C8.22123 5.0625 7.45996 5.29343 6.81244 5.72609C6.16492 6.15875 5.66024 6.7737 5.36222 7.49319C5.0642 8.21267 4.98623 9.00437 5.13816 9.76817C5.29009 10.532 5.6651 11.2336 6.21577 11.7842C6.76643 12.3349 7.46803 12.7099 8.23183 12.8618C8.99563 13.0138 9.78733 12.9358 10.5068 12.6378C11.2263 12.3398 11.8412 11.8351 12.2739 11.1876C12.7066 10.54 12.9375 9.77877 12.9375 9C12.9375 7.95571 12.5226 6.95419 11.7842 6.21577C11.0458 5.47734 10.0443 5.0625 9 5.0625Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	const size24 = (
-		<svg
-			width='24'
-			height='24'
-			viewBox='0 0 24 24'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M16.3748 3.24984C17.5342 3.25331 18.6451 3.71539 19.4648 4.53517C20.2846 5.35495 20.7467 6.46582 20.7502 7.62516V16.3748C20.7467 17.5342 20.2846 18.6451 19.4648 19.4648C18.6451 20.2846 17.5342 20.7467 16.3748 20.7502H7.62516C6.46582 20.7467 5.35495 20.2846 4.53517 19.4648C3.71539 18.6451 3.25331 17.5342 3.24984 16.3748V7.62516C3.25331 6.46582 3.71539 5.35495 4.53517 4.53517C5.35495 3.71539 6.46582 3.25331 7.62516 3.24984H16.3748ZM16.3748 1.5H7.62516C4.25625 1.5 1.5 4.25625 1.5 7.62516V16.3748C1.5 19.7437 4.25625 22.5 7.62516 22.5H16.3748C19.7437 22.5 22.5 19.7437 22.5 16.3748V7.62516C22.5 4.25625 19.7437 1.5 16.3748 1.5Z'
-				fill={color}
-			/>
-			<path
-				d='M17.6873 7.62516C17.4277 7.62516 17.174 7.54818 16.9581 7.40396C16.7423 7.25974 16.5741 7.05476 16.4747 6.81493C16.3754 6.5751 16.3494 6.3112 16.4001 6.0566C16.4507 5.802 16.5757 5.56814 16.7593 5.38458C16.9428 5.20102 17.1767 5.07602 17.4313 5.02538C17.6859 4.97473 17.9498 5.00072 18.1896 5.10006C18.4294 5.1994 18.6344 5.36763 18.7786 5.58347C18.9229 5.79931 18.9998 6.05307 18.9998 6.31266C19.0002 6.48512 18.9665 6.65596 18.9007 6.81537C18.8348 6.97477 18.7382 7.11961 18.6162 7.24156C18.4943 7.36351 18.3494 7.46018 18.19 7.526C18.0306 7.59183 17.8598 7.62553 17.6873 7.62516ZM12 8.49985C12.6923 8.49985 13.369 8.70513 13.9446 9.08973C14.5202 9.47433 14.9688 10.021 15.2337 10.6606C15.4986 11.3001 15.5679 12.0039 15.4329 12.6829C15.2978 13.3618 14.9645 13.9855 14.475 14.475C13.9855 14.9645 13.3618 15.2979 12.6828 15.4329C12.0039 15.568 11.3001 15.4986 10.6605 15.2337C10.021 14.9688 9.47433 14.5202 9.08972 13.9446C8.70512 13.369 8.49984 12.6923 8.49984 12C8.50083 11.072 8.86992 10.1823 9.52611 9.52612C10.1823 8.86992 11.072 8.50084 12 8.49985ZM12 6.75C10.9616 6.75 9.94661 7.05791 9.08325 7.63479C8.2199 8.21166 7.54699 9.0316 7.14963 9.99092C6.75227 10.9502 6.64831 12.0058 6.85088 13.0242C7.05345 14.0426 7.55346 14.9781 8.28769 15.7123C9.02191 16.4465 9.95737 16.9466 10.9758 17.1491C11.9942 17.3517 13.0498 17.2477 14.0091 16.8504C14.9684 16.453 15.7883 15.7801 16.3652 14.9167C16.9421 14.0534 17.25 13.0384 17.25 12C17.25 10.6076 16.6969 9.27226 15.7123 8.28769C14.7277 7.30313 13.3924 6.75 12 6.75Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	switch (size.toString()) {
-		case '16':
-			return size18;
-		case '24':
-			return size24;
-		default:
-			return size18;
-	}
-};
