+++ src/components/social-icons/Instagram/Instagram18.tsx
@@ -1,0 +1,24 @@
+import React, { FC } from 'react';
+import { ISocialIconProps } from '../type';
+
+export const IconInstagram18: FC<ISocialIconProps> = ({
+	size = 18,
+	color = 'currentColor',
+}) => (
+	<svg
+		width={size}
+		height={size}
+		viewBox='0 0 18 18'
+		fill='none'
+		xmlns='http://www.w3.org/2000/svg'
+	>
+		<path
+			d='M12.2811 2.43738C13.1506 2.43998 13.9838 2.78654 14.5986 3.40138C15.2135 4.01621 15.56 4.84936 15.5626 5.71887V12.2811C15.56 13.1506 15.2135 13.9838 14.5986 14.5986C13.9838 15.2135 13.1506 15.56 12.2811 15.5626H5.71887C4.84936 15.56 4.01621 15.2135 3.40138 14.5986C2.78654 13.9838 2.43998 13.1506 2.43738 12.2811V5.71887C2.43998 4.84936 2.78654 4.01621 3.40138 3.40138C4.01621 2.78654 4.84936 2.43998 5.71887 2.43738H12.2811ZM12.2811 1.125H5.71887C3.19219 1.125 1.125 3.19219 1.125 5.71887V12.2811C1.125 14.8078 3.19219 16.875 5.71887 16.875H12.2811C14.8078 16.875 16.875 14.8078 16.875 12.2811V5.71887C16.875 3.19219 14.8078 1.125 12.2811 1.125Z'
+			fill={color}
+		/>
+		<path
+			d='M13.2655 5.71887C13.0708 5.71887 12.8805 5.66114 12.7186 5.55297C12.5567 5.44481 12.4306 5.29107 12.3561 5.1112C12.2816 4.93133 12.2621 4.7334 12.3 4.54245C12.338 4.3515 12.4318 4.1761 12.5694 4.03843C12.7071 3.90077 12.8825 3.80701 13.0735 3.76903C13.2644 3.73105 13.4623 3.75054 13.6422 3.82505C13.8221 3.89955 13.9758 4.02572 14.084 4.1876C14.1921 4.34948 14.2499 4.5398 14.2499 4.73449C14.2501 4.86384 14.2249 4.99197 14.1755 5.11153C14.1261 5.23108 14.0536 5.33971 13.9622 5.43117C13.8707 5.52263 13.7621 5.59513 13.6425 5.6445C13.523 5.69387 13.3948 5.71915 13.2655 5.71887ZM9 6.37488C9.51919 6.37488 10.0267 6.52885 10.4584 6.8173C10.8901 7.10575 11.2266 7.51574 11.4253 7.99541C11.624 8.47509 11.676 9.00291 11.5747 9.51214C11.4734 10.0214 11.2234 10.4891 10.8562 10.8562C10.4891 11.2234 10.0214 11.4734 9.51213 11.5747C9.00291 11.676 8.47509 11.624 7.99541 11.4253C7.51573 11.2266 7.10574 10.8901 6.81729 10.4584C6.52884 10.0267 6.37488 9.5192 6.37488 9C6.37563 8.30401 6.65244 7.63673 7.14458 7.14459C7.63672 6.65244 8.304 6.37563 9 6.37488ZM9 5.0625C8.22123 5.0625 7.45996 5.29343 6.81244 5.72609C6.16492 6.15875 5.66024 6.7737 5.36222 7.49319C5.0642 8.21267 4.98623 9.00437 5.13816 9.76817C5.29009 10.532 5.6651 11.2336 6.21577 11.7842C6.76643 12.3349 7.46803 12.7099 8.23183 12.8618C8.99563 13.0138 9.78733 12.9358 10.5068 12.6378C11.2263 12.3398 11.8412 11.8351 12.2739 11.1876C12.7066 10.54 12.9375 9.77877 12.9375 9C12.9375 7.95571 12.5226 6.95419 11.7842 6.21577C11.0458 5.47734 10.0443 5.0625 9 5.0625Z'
+			fill={color}
+		/>
+	</svg>
+);
